Creating the discord app, adding bot to our discord server, logging the bot in
I created a server in the discord website and called it discord bot tutorial, then I went to the url add the client id and scope with bot at the end to tell the site that i want to create a bot, i logged in to the bot. I setup nodemon in the package.json so all i have to do from now on is type npm start.